# saki-controlplane/internal/gen/openapi/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package openapi

import (
	"fmt"
)

func (s *ErrorResponseStatusCode) Error() string {
	return fmt.Sprintf("code %d: %+v", s.StatusCode, s.Response)
}

// Ref: #/components/schemas/AuthTokenResponse
type AuthTokenResponse struct {
	Token       string   `json:"token"`
	UserID      string   `json:"user_id"`
	Permissions []string `json:"permissions"`
}

// GetToken returns the value of Token.
func (s *AuthTokenResponse) GetToken() string {
	return s.Token
}

// GetUserID returns the value of UserID.
func (s *AuthTokenResponse) GetUserID() string {
	return s.UserID
}

// GetPermissions returns the value of Permissions.
func (s *AuthTokenResponse) GetPermissions() []string {
	return s.Permissions
}

// SetToken sets the value of Token.
func (s *AuthTokenResponse) SetToken(val string) {
	s.Token = val
}

// SetUserID sets the value of UserID.
func (s *AuthTokenResponse) SetUserID(val string) {
	s.UserID = val
}

// SetPermissions sets the value of Permissions.
func (s *AuthTokenResponse) SetPermissions(val []string) {
	s.Permissions = val
}

// Ref: #/components/schemas/CurrentUserResponse
type CurrentUserResponse struct {
	UserID      string   `json:"user_id"`
	Permissions []string `json:"permissions"`
}

// GetUserID returns the value of UserID.
func (s *CurrentUserResponse) GetUserID() string {
	return s.UserID
}

// GetPermissions returns the value of Permissions.
func (s *CurrentUserResponse) GetPermissions() []string {
	return s.Permissions
}

// SetUserID sets the value of UserID.
func (s *CurrentUserResponse) SetUserID(val string) {
	s.UserID = val
}

// SetPermissions sets the value of Permissions.
func (s *CurrentUserResponse) SetPermissions(val []string) {
	s.Permissions = val
}

// Ref: #/components/schemas/ErrorResponse
type ErrorResponse struct {
	Code    string `json:"code"`
	Message string `json:"message"`
}

// GetCode returns the value of Code.
func (s *ErrorResponse) GetCode() string {
	return s.Code
}

// GetMessage returns the value of Message.
func (s *ErrorResponse) GetMessage() string {
	return s.Message
}

// SetCode sets the value of Code.
func (s *ErrorResponse) SetCode(val string) {
	s.Code = val
}

// SetMessage sets the value of Message.
func (s *ErrorResponse) SetMessage(val string) {
	s.Message = val
}

// ErrorResponseStatusCode wraps ErrorResponse with StatusCode.
type ErrorResponseStatusCode struct {
	StatusCode int
	Response   ErrorResponse
}

// GetStatusCode returns the value of StatusCode.
func (s *ErrorResponseStatusCode) GetStatusCode() int {
	return s.StatusCode
}

// GetResponse returns the value of Response.
func (s *ErrorResponseStatusCode) GetResponse() ErrorResponse {
	return s.Response
}

// SetStatusCode sets the value of StatusCode.
func (s *ErrorResponseStatusCode) SetStatusCode(val int) {
	s.StatusCode = val
}

// SetResponse sets the value of Response.
func (s *ErrorResponseStatusCode) SetResponse(val ErrorResponse) {
	s.Response = val
}

// Ref: #/components/schemas/HealthResponse
type HealthResponse struct {
	Status string `json:"status"`
}

// GetStatus returns the value of Status.
func (s *HealthResponse) GetStatus() string {
	return s.Status
}

// SetStatus sets the value of Status.
func (s *HealthResponse) SetStatus(val string) {
	s.Status = val
}

// Ref: #/components/schemas/LoginRequest
type LoginRequest struct {
	UserID      string   `json:"user_id"`
	Permissions []string `json:"permissions"`
}

// GetUserID returns the value of UserID.
func (s *LoginRequest) GetUserID() string {
	return s.UserID
}

// GetPermissions returns the value of Permissions.
func (s *LoginRequest) GetPermissions() []string {
	return s.Permissions
}

// SetUserID sets the value of UserID.
func (s *LoginRequest) SetUserID(val string) {
	s.UserID = val
}

// SetPermissions sets the value of Permissions.
func (s *LoginRequest) SetPermissions(val []string) {
	s.Permissions = val
}

// Ref: #/components/schemas/Project
type Project struct {
	ID   string `json:"id"`
	Name string `json:"name"`
}

// GetID returns the value of ID.
func (s *Project) GetID() string {
	return s.ID
}

// GetName returns the value of Name.
func (s *Project) GetName() string {
	return s.Name
}

// SetID sets the value of ID.
func (s *Project) SetID(val string) {
	s.ID = val
}

// SetName sets the value of Name.
func (s *Project) SetName(val string) {
	s.Name = val
}

// RequirePermissionNoContent is response for RequirePermission operation.
type RequirePermissionNoContent struct{}